Updates Projects Controller and Service with the new relation
- In ProjectsConetroller: Now create method receives the id of the user that will be owner of the project,  this will be necesary to give 'owner' access level to this user. Also, adds a new method 'addUserToProject', that will try to add an user to an existing project.

- In ProjectsService:
Create method now needs to start a transaction, because when creating a new project, two things must happen together: First, saving the project in the databse, and second, create the relation of the user that is creating the project. If one of these things fails, could mean that the database stores wrong information .

Also adds the 'addUserToProject' method, that first will verify if the user already exists in the project (throwing an error in that case), and then save the userProject relation in the users_project table.