Adds unit tests to UsersService
Adds unit tests for every method in UsersService. This involves creating a testing module that utilizes an in-memory SQLite database to simulate real scenarios and test each method,  for example, creating an user and finging him using the autogenerated id, as well as verifying that the methods correctly handle error conditions.