Updates UsersServices tests. Adds unit testing to TasksService
- UsersService:
Changes all error tests to now throw the exact error message, making each test more specific. This change was made because only verifying that an error was being thrown was very generic, and it was possible that the expected error was not being thrown.

Also, updates the update test to now send a uuidSample and an object with the edited data to avoid NestJS pipes validation.

-TasksService
Adds unit tests for every method in TasksService. This involves creating a testing module that utilizes an in-memory SQLite database to simulate real scenarios and test each method. The real TasksService doesn’t directly use projects or users repositories, but it was necessary for the testing module to create relationships between task-project or task-user, achieving tests that verify the real functionality of the service.