Adds unit tests to ProjectsService
Adds unit tests for every method in ProjectsService. This involves creating a testing module that utilizes an in-memory SQLite database to simulate real scenarios and test each method. The real ProjectsServcie doesn’t directly use UserRepository, but it was necessary for the testing module to create relationships between user-project, achieving tests that verify the real functionality of the service.