Adds unit tests to ErrorManager and Updates ErrorManager
Adds unit tests to ErrorManager verifying if its 'createSignatureError' method works properly and if the errors thrown are instances of HttpException.

Also, updates ErrorManager, removing a default error on line 15. It was impossible for this to happen since the new ErrorManager only accepts error names from the keyof typeof HttpStatus.