Front: Updates the signup form to be able to show or not the password
Adds a state called 'showPassword', which is a boolean and is used to show or not the data entered in the 'password' and 'confirm password' components, if its true the input type changes to 'text', if not, it returns to 'password'.

Also adds a handler called 'handleClickShowPassword' to change the state of showPassword when the users clicks the 'show password' icon.