Front: Adds a NotificationProvider
Adds a notification provider that will wrap the App component and return the Notification component with the props defined by any of the get Success/Error/Warning/Info functions.

These four functions are returned by the useNotification custom hook, to allowing the Notification component to be displayed from anywhere in the app. This is achieved by using the useNotification hook and calling one of the get Success/Error/Warning/Info functions.