Front: Adds an auth thunk and new reducer cases in AuthSlice
Introduces an authentication thunk, that attempts to authenticate the user and retrieve their access token using the `auth.login` method to communicate with the backend. If it fails, returns the error with the rejectWithValue function from redux.

Additionally, enhances the `initialState` variable to include comprehensive state information related to the authentication process.

Furthermore, adds three new cases in `extraReducers` within the authentication slice (`AuthSlice`):
- Pending: Handles pending authentication requests.
- Fulfilled: Manages successful authentication .
- Rejected: Sets the failed authentication.

Each state updates the auth state with the corresponding and necessary information.