Front: Adds session persistence with access token cookie
- In RouterLayout.tsx:
Adds effects to set cookies when the accessToken value from the auth reducer changes (i.e., when the user logs in) and another effect to remove the cookie if it has expired.

Changes the way the component handles error notifications; it now uses an effect.

- In NavBar.tsx:
Replaces the 'login' and 'register' buttons with a single 'logout' button. The navbar is now only displayed when the user is authenticated (i.e., logged in), making it unnecessary to show the login and register buttons.

The logout button removes the access token cookie, sets the isAuth value from the authReducer to false using the 'logout' reducer, and redirects the user to the '/login' route.

Updates styles for the app bar by adding a box shadow and enhances the 'Projects Manager' typography with better font styling.