Front: Adds an initial HomePage
Implements an inital home page, which will obtain the complete user data with the custom hook called 'useHomePage', using the getUserReducer with the user id and access token obtained by the authReducer state.

With this information, the HomePage component will conditional render a loading component and then a welcome message, followed by either a list of the project that the user is part of, or a message saying that the user doesnt belong to any project.