Front: Adds the user projects list in HomePage component
Creates a new component called 'Project' that represent a project accordion in the projects list in the home page.

Depending of the user access level in the clicked project, it will display operation buttons with popovers that contains help texts.

The states setted in the Project component are necessary for showing the right popover during hover and to remove it on mouse leave

In the useHome custom hook, the expand state and the handleAccordionChange are necessary to allow the user to click just one project at the time.