Front: Adds a basic modal component to create a project
Implements a new component called 'CreateProjectModal' that will be  rendered in the home page, being a button that displays a modal that works as a form, with two inputs: one containing the project title and the other the project description.

The modal works as a form, this means that it needs to be validated, to doing so, implements a formik and yup validation system in the new custom hook called 'useCreateProject', where the modal 'open' state and its open-close handlers will be contained.

Finally, adds a new type called CreateProjectType for the modal form and an interface for the new custom useCreateProject hook