Front: Refactores the 'HomePage' component with better organization
Creates a new component to display a message when the user doesnt belong to any project, depending of its role, will show either a button to create a new project, or a message saying that with 'BASIC' role a project cannot be created.

Also, rename the 'projectsList' variable to 'renderProjectsList' to better reflect what it does.