Front: Adds functionality to create a project
- In projects.api.ts:
Adds a new method to connect to the backend and create a project, sending the access token and the post body into the header, throwing an error if the operation fails.

- In useCreateProject.ts:
First, obtains the userData and access token using the authReducer. When the form is submitted, it triggers the creation of the project by sending the user ID, values entered in the form, and the access token.

Also, creates a new state called 'loading', which is used to set the 'loading' property in the new CreateProjectModal component: 'LoadingButton', disabling the button and displaying a loading spinner when the user submit the form.

Finally, uses the useNotification provider to display a success notification, close the modal and navigate the user to 'login' route, this is necessary to force a HomePage component re-rendering, obtaining the updated projects with the new one (the user is already logged, meaning that the LoginPage component wont be rendered, instead, a Navigate react component to '/' route will be).