Front: Adds functionality to delete a project
- In projects.api.ts:
Adds a new method to connect to the backend and delete a project, sending the access token into the header, throwing an error if the operation fails.

- In useDeleteProject.ts:
First, obtains the access token using the authReducer. When the 'confirm delete' modal button is clicked,  triggers the delete of the project usign the new projects.delete method, sending the project id and access token.

Also, creates a state called 'loading', which is used to set the 'loading' property in the LoadingButton component of the DeleteModal, disabling the button and displaying a loading spinner when the user confirms the delete.

Finally, uses the useNotification provider to display a success notification and navigate the user to 'login' route, this is necessary to force a HomePage component re-rendering, obtaining the updated projects without the deleted one (the user is already logged, meaning that the LoginPage component wont be rendered, instead, a Navigate react component to '/' route will be).