Front: Updates the useCreateProject hook with edit operation
- In projects.api.ts:
Adds a new method to connect to the backend and edit a project, sending the new project data and the access token into the header, throwing an error if the operation fails.

- In useSaveProject:
Refactor the useCreateProject custom hook to be used also to edit a project, now it will receive initial params, that by default will be empty or null, meaning that if the hook is called with no params, the operation is a creation, otherwhise, it will be an edition. Also, changes the names of the returned variables to be  more specific.

In both scenarios, the onSubmit handler will try to connect to the backend sending the data submitted in the form, and then notificate the user a success operation, redirecting to '/login' route (this is necessary the user is already logged, meaning that the LoginPage component wont be rendered, instead, a Navigate react component to '/' route will be). If this fails, it will display a notification error to user

- In index.tsx:
Implements the new updated useSaveProject hook

-In custom.hooks.interface.ts
Updates the interface with the new names of the variables