Front: Updates the DeleteModal component to be a controlled component
- In DeleteModal/index.tsx:
Removes the use of the useDeleteProject custom hook.  It will now receive this properties by its father component using the that custom hook, making it a controlled component.

- In useProjectsList.ts:
This custom hook will no longer define the 'open' state of the DeleteModal.

- In useDeleteProjects.ts
Implements the 'open' state of the DeleteModal removed in the useProjectsList custom hook, also changes the name of 'loading' state to be more specific.

- In ProjectListElement/index.tsx
Implements the useDeleteProject custom hook, sending the received properties to DeleteModal component as props.