Front: Adds a function for adding an user to a project
This function connects to the backend by a 'POST' request, it tryes to add an user to a project, sending the access token for authorization, and a body, which must contain an user id and access level to asign to that user.

If this success, returns the JSON response, otherwise, throws the received error.