Front: Adds a 'add user to project' form type inside a modal
Implements a new modal that will contain a form to add an user to a project, using the 'useAddUserToProject' custom hook.

For doign it, adds an 'Autocomplete' mui component, that renders a 'TextField' input, which allows partial users searching by the username entered. These two components will contain the same formik value, allowing to submit the form either typing the exact username in the 'TextField', or selecting one from the list inside of the 'Autocomplete'. To achieve this, is necessary to update the value of formik.value.user when the value of the 'Autocomplete' or 'TextField' changes.

The validation and feedback of the form will be setted by a formik object received by the 'useAddUserToProject' custom hook.