Front: Adds a new route to render a project's data
Adds a new route in the Router, that will render the new component called 'ProjectPage'. This component will display information about a project  (which is found by its id usign the 'useProject' custom hook) using another new component called 'ProjectStats'.

Also, completes the ProjectData type in the project.type.ts, adding the users and tasks types and creating a new type for the users involved in a project