Front: Refactor ProjectStats component
Changes the conditional rendering of the TasksStats component. The section that informs about the project not having any task is no longer displayed in this component, which now render the stats only, simulating an if using '&&' operator, depending if the project does have tasks created.

Now, the message and button that informs about the project not having any task is inside ProjectStats component.