Front: Adds a modal and a custom hook to save tasks
Creates a new modal for saving tasks, containing a form with its title, description and status. This form will be handled by usign a formik object defined inside 'useSaveTask' custom hook.

The 'useSaveTask' custom hook will set the state that defines if the 'SaveTaskModal' is open or not, and handlers to open/close it. Finally, this hook will be used either to save a new task, or edit an existing one, depending if its called with params or not.