Front: Adds functionality to saving tasks
- in tasks.api.ts:
Creates two methods to the backend by a 'POST' and 'PUT' requests, it tries to create/edit a task insde a project, sending the access token for authorization, and a body, which contains a task title, description and status.

If this success, returns the JSON response, otherwise, throws the received error.

- In validateForm.ts:
Adds a schema to validate 'save task' form

- In useSaveTask.ts:
Adds the new validation schema created to the formik object, and functionality while submitting the form. First, it will disable the 'add' button by setting the loadingSaveTaskButton into true, and then will try to edit or create (depending if the hook received a task id) a task in a project, using the notification provider to notificate the success operation. If this fails, uses the notification provider to display the message and set the 'loadingSaveTaskButton' into false again.