Front: Adds an 'add task' operation in the TasksPage component
Implements a new component in the TasksPage component called 'AddTasskToProject' (which is rendered only if the user has the required role), that will contain a button that triggers a modal with a form, allowing an user to create a task in a project.

In order for doing this, makes use of the 'useSaveTask' custom hook to handle validation, open/close and submitting of the form modal. Also, the hook now updates the getTasksReducer, to force a re-render in the TasksDataGrid component, achieving a real time update. The navigate(/project/tasks/${projectId}), is necessary in the case where the user is creating a task in the ProjectPage component, redirecting the user to that route and rendering the TasksDataGrid, achieving real time update.

Finally, adds a missing type and interface in the task.type.ts and redux.state.interface.ts, respectively.