Front: Fixes a bug where add task to project button where not being displayed
The button was not being shown because the comparission must be between access levels and not roles.

In order to fix this, it is necessary to obtain the access level of the user that want to create a task in a project. With that in mind, the 'useTaskPage' custom hook now makes uses of 'getUserReducer ' and 'getUserThunk' from redux, to obtain the user and find the necessary access level