Front: Adds functionality for deleting a task
Creates a new custom hook called 'useDeleteTask' that will set the necessary states to handle the delete modal, like states to open/close it, or to disable the 'confirm delete' button after click.

Also, creates a function that will execute the delete operation, that will try to delete the task with the received task id, updating the tasks list with the 'getTasksThunk' reducer, forcing a re-render in the TaskPage component, achieving changes in real time. After that, a notification will be displayed with the notification provider.

Finally, implements a new method in the tasks.api.ts to connect to the backend by 'DELETE' method, sending the access token as a header for authentication (and verify if the user has te necessary role to execute the operation), and trying to delete the task with the received id.